Nearly Complete First Draft
I have the major components Burch wanted, I need one or two more major
components to bulk it up to the minimum page requirement. Maybe we can
add in some more about what we each bring and how we could compensate
if one were to drop. Also, the attachments need to be added.