WebCrawler
A new function has been added. This is the beginnings of a potential
function to search google scholar to construct the article network we
would need to analyze the entire 511 network’s research. I think this
would be too large to implement, especially since we are mainly
focusing on the algorithm.

-CM